test(ReservationCreationRequest): DTO 검증 테스트코드 수정
