--- src/test/java/roomescape/domain/reservation/dto/CreateReservationRequestTest.java
@@ -1,89 +1,0 @@
-package roomescape.domain.reservation.dto;
-
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
-import org.junit.jupiter.api.Test;
-import roomescape.support.exception.RoomescapeException;
-
-class CreateReservationRequestTest {
-
-    @Test
-    void 이름이_null이면_예외가_발생한다() {
-        // given
-        ReservationCreationRequest request = new ReservationCreationRequest(
-            null,
-            1L,
-            1L,
-            1L
-        );
-
-        // when & then
-        assertThatThrownBy(request::validate)
-            .isInstanceOf(RoomescapeException.class)
-            .hasMessage("이름은 비어 있을 수 없습니다.");
-    }
-
-    @Test
-    void 이름이_공백이면_예외가_발생한다() {
-        // given
-        ReservationCreationRequest request = new ReservationCreationRequest(
-            "   ",
-            1L,
-            1L,
-            1L
-        );
-
-        // when & then
-        assertThatThrownBy(request::validate)
-            .isInstanceOf(RoomescapeException.class)
-            .hasMessage("이름은 비어 있을 수 없습니다.");
-    }
-
-    @Test
-    void 날짜가_null이면_예외가_발생한다() {
-        // given
-        ReservationCreationRequest request = new ReservationCreationRequest(
-            "보예",
-            null,
-            1L,
-            1L
-        );
-
-        // when & then
-        assertThatThrownBy(request::validate)
-            .isInstanceOf(RoomescapeException.class)
-            .hasMessage("날짜는 필수입니다.");
-    }
-
-    @Test
-    void 시간_id가_null이면_예외가_발생한다() {
-        // given
-        ReservationCreationRequest request = new ReservationCreationRequest(
-            "보예",
-            1L,
-            null,
-            1L
-        );
-
-        // when & then
-        assertThatThrownBy(request::validate)
-            .isInstanceOf(RoomescapeException.class)
-            .hasMessage("시간은 필수입니다.");
-    }
-
-    @Test
-    void 테마_id가_null이면_예외가_발생한다() {
-        // given
-        ReservationCreationRequest request = new ReservationCreationRequest(
-            "보예",
-            1L,
-            1L,
-            null
-        );
-
-        // when & then
-        assertThatThrownBy(request::validate)
-            .isInstanceOf(RoomescapeException.class)
-            .hasMessage("테마는 필수입니다.");
-    }
-}
